requirements.txt
Django~=4.0.3
django-widget-tweaks~=1.4.12
django-allauth~=0.49.0
Pillow~=9.0.1
django-stdimage~=5.3.0
django-markdownx~=3.0.1
django-mdeditor~=0.1.20
fontawesomefree~=6.1.2
Markdown~=3.3.7
django-cleanup~=6.0.0